[maximal_planar] move Phase2 to own file
